refactor(newrelic): Break newrelic (cmd) commands out to specific files
